+++ ElvUI_MerathilisUI/Core/Changelog/7.07.lua
@@ -1,8 +1,13 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[707] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "27.01.2026",
+	FIXES = {
+		"[Portraits]: Fixed some copy-paste errors in the code.",
+		"[UnitFrames]: Fixed an error with castbars",
+		"[UnitFrames]: Remove the Tag [name:gradient] for now and replace it with [name:abbrev:medium]",
+		"              you need to do the install for NamePlates & UnitFrames again to apply the changes.",
+	},
 	NEW = {},
 	IMPROVEMENTS = {},
 }